requirements.txt
Flask==3.1.1
Flask-Mail==0.10.0
Werkzeug==3.1.3
email-validator==2.2.0
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.31.0